feat(rollout): add possibility to fetch binary from local filesystem (nodejs)
