+++ packages/rollout/tsconfig.lib.json
@@ -13,7 +13,7 @@
         "removeComments": true,
         "types": ["jest", "node"]
     },
-    "include": ["./src/index.ts"],
+    "include": ["./src/index.ts", "./src/utils/fetch-browser.ts"],
     "exclude": [
         "node_modules",
         "**/__tests__/**",
